Extract Box, imread
